renamed question1.py in question2 to utils.py
