+++ solution/question2/utils.py
